I added two lines of code:
First line is in the Manifest which is commeted and starts by Ibrahim. This line allows the app to connect to the internet.
Second line is in the app\build.gradle. This line for importing the okhttp library that gives us access to avery handy API for building and sending HTTP requests.