I modified LoginActivity and SignUpActivity
I added addition method that connect to the server and check the response and deal with it. In the LoginActivity it ckecks if the login success. It takes the user to the SearchActivity otherwise it showa a toast message to the user.
In the SignupActivity it check if the signup success. either take the user to the login page, in sucess, or show a toast message.